requirements.txt
psycopg2-binary
fastapi
uvicorn
transformers
twilio
streamlit
python-dotenv